+++ boarding/src/main/java/ca/ulaval/glo4002/thunderbird/boarding/domain/baggage/collection/PersonalBaggageCollection.java
@@ -40,7 +40,7 @@ public class PersonalBaggageCollection extends BaggageCollection {
     }
 
     @Override
-    public float calculateTotalCost() {
+    public double calculateTotalCost() {
         return 0;
     }
