+++ boarding/src/test/java/ca/ulaval/glo4002/thunderbird/boarding/domain/baggage/collection/PersonalBaggageCollectionTest.java
@@ -21,7 +21,7 @@ public class PersonalBaggageCollectionTest {
     private static final float DELTA = 0.01f;
     private static final String TYPE = "personal";
     private static final List<Baggage> EMPTY_LIST = new ArrayList<>();
-    private static final float BAGGAGE_TOTAL_COST = 0;
+    private static final double BAGGAGE_TOTAL_COST = 0;
 
     private PersonalBaggageCollection baggageCollection;
     private Baggage baggage;
@@ -40,7 +40,7 @@ public class PersonalBaggageCollectionTest {
 
     @Test
     public void whenCalculatingTotalCost_shouldReturnFree() {
-        float cost = baggageCollection.calculateTotalCost();
+        double cost = baggageCollection.calculateTotalCost();
 
         assertEquals(BAGGAGE_TOTAL_COST, cost, DELTA);
     }
