+++ boarding/src/test/java/ca/ulaval/glo4002/thunderbird/boarding/domain/baggage/collection/StandardBaggageCollectionTest.java
@@ -18,10 +18,10 @@ import static org.mockito.Mockito.mock;
 
 public class StandardBaggageCollectionTest {
 
-    private static final float DELTA = 0.01f;
+    private static final double DELTA = 0.01f;
     private static final String TYPE = "standard";
     private static final List<Baggage> EMPTY_LIST = new ArrayList<>();
-    private static final float BAGGAGE_TOTAL_COST = 0;
+    private static final double BAGGAGE_TOTAL_COST = 0;
 
     private StandardBaggageCollection baggageCollection;
     private Baggage baggage;
@@ -40,7 +40,7 @@ public class StandardBaggageCollectionTest {
 
     @Test
     public void whenCalculatingTotalCost_shouldReturnFree() {
-        float cost = baggageCollection.calculateTotalCost();
+        double cost = baggageCollection.calculateTotalCost();
 
         assertEquals(BAGGAGE_TOTAL_COST, cost, DELTA);
     }
